Use weakdeps on Julia v1.9
