+++ src/StructArrays.jl
@@ -12,8 +12,6 @@ include("utils.jl")
 include("collect.jl")
 include("sort.jl")
 include("lazy.jl")
-include("tables.jl")
-include("staticarrays_support.jl")
 
 # Implement refarray and refvalue to deal with pooled arrays and weakrefstrings effectively
 import DataAPI: refarray, refvalue
@@ -29,15 +27,10 @@ end
 import Adapt
 Adapt.adapt_structure(to, s::StructArray) = replace_storage(x->Adapt.adapt(to, x), s)
 
-# for GPU broadcast
-import GPUArraysCore
-function GPUArraysCore.backend(::Type{T}) where {T<:StructArray}
-    backends = map_params(GPUArraysCore.backend, array_types(T))
-    backend, others = backends[1], tail(backends)
-    isconsistent = mapfoldl(isequal(backend), &, others; init=true)
-    isconsistent || throw(ArgumentError("all component arrays must have the same GPU backend"))
-    return backend
+@static if !isdefined(Base, :get_extension)
+    include("../ext/StructArraysGPUArraysCoreExt.jl")
+    include("../ext/StructArraysTablesExt.jl")
+    include("../ext/StructArraysStaticArraysCoreExt.jl")
 end
-always_struct_broadcast(::GPUArraysCore.AbstractGPUArrayStyle) = true
 
 end # module
